refactor: Keep visibility schema in list-workflows, move helpers to utils
Keep visibilityQuerySchema inline in list-workflows (exported for
count-workflows to consume). Move getVisibilityQuery and
validateArchivedQueryParams to src/utils/visibility/. Remove
route-handlers/shared/.
