+++ src/route-handlers/list-workflows/schemas/list-workflows-query-params-schema.ts
@@ -1,7 +1,42 @@
 import { z } from 'zod';
 
-import validateArchivedQueryParams from '@/route-handlers/shared/helpers/validate-archived-query-params';
-import visibilityQuerySchema from '@/route-handlers/shared/schemas/visibility-query-schema';
+import { SORT_ORDERS } from '@/utils/sort-by';
+import validateArchivedQueryParams from '@/utils/visibility/validate-archived-query-params';
+import isWorkflowStatus from '@/views/shared/workflow-status-tag/helpers/is-workflow-status';
+import { type WorkflowStatus } from '@/views/shared/workflow-status-tag/workflow-status-tag.types';
+
+const workflowStatusSchema = z.custom<WorkflowStatus>(isWorkflowStatus, {
+  message: 'Invalid workflow status',
+});
+
+export const visibilityQuerySchema = z.object({
+  listType: z.enum(['default', 'archived']),
+  inputType: z.enum(['search', 'query']),
+  search: z
+    .string()
+    .trim()
+    .optional()
+    .transform((search) =>
+      search === undefined
+        ? undefined
+        : search.replace(/['"]/g, (match) => `\\${match}`)
+    ),
+  query: z.string().optional(),
+  statuses: z
+    .union([
+      workflowStatusSchema.transform((status) => [status]),
+      z.array(workflowStatusSchema),
+    ])
+    .optional(),
+  timeColumn: z
+    .enum(['StartTime', 'CloseTime'])
+    .optional()
+    .default('StartTime'),
+  timeRangeStart: z.string().datetime().optional(),
+  timeRangeEnd: z.string().datetime().optional(),
+  sortColumn: z.string().optional(),
+  sortOrder: z.enum(SORT_ORDERS, { message: 'Invalid sort order' }).optional(),
+});
 
 const listWorkflowsQueryParamSchema = visibilityQuerySchema
   .merge(
